+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRB2DYSQ1JBYT4EXSC8ZR3RS"
+        "01KRB99XAYP355YY0D5GWYBA18"
     )
     r"""The unique identifier of the data source"""