+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRB2DYSHJ3X614YFZTV9VF97"
+        "01KRB99XA7MRKF1KDKCS93Z4X1"
     )
     r"""The unique identifier of the data source"""