+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRB2DYSN18WGTHPCZ5GWP9CF"
+        "01KRB99XAJ5T50KH70XD7VJYHF"
     )
     r"""The unique identifier of the data source"""