+++ packages/orq-rc/RELEASES.md
@@ -7719,3 +7719,13 @@ Based on:
 - [python v4.9.0-rc.39] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.39 - packages/orq-rc
+
+## 2026-05-11 10:32:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.40] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.40 - packages/orq-rc